remove workspace

+++ src/allocator.rs
@@ -261,7 +261,7 @@ impl Allocator {
 
     /// Return a ptr given a shareable offset - calculated by `offset`.
     pub fn ptr_from_offset(&self, offset: usize) -> NonNull<u8> {
-        unsafe { self.header.byte_add(offset as usize) }.cast()
+        unsafe { self.header.byte_add(offset) }.cast()
     }
 
     /// Find the slab index and index within the slab for a given offset.
